Added prompt to confirm deletion for project, workflow, execution, files & fixed hard coded prompt  (#1580)
Addresses #1574
This PR adds a confirmation prompt displaying the type of deleted item(project/workflow/execution/file) and the name of the item.

Before:
![before](https://user-images.githubusercontent.com/44384988/178089850-d425b11f-af1a-4dbc-b203-79f6e8ff7f1b.gif)

After:
![after](https://user-images.githubusercontent.com/44384988/178090251-49c37790-0d26-4048-8346-9565b77bc89a.gif)

+++ core/new-gui/src/app/dashboard/component/feature-container/user-file-section/user-file-section.component.ts
@@ -11,6 +11,8 @@ import { NotificationService } from "../../../../common/service/notification/not
 import { UserProjectService } from "../../../service/user-project/user-project.service";
 import { UserProject } from "../../../type/user-project";
 import Fuse from "fuse.js";
+import { DeletePromptComponent } from "../../delete-prompt/delete-prompt.component";
+import { from } from "rxjs";
 
 export const ROUTER_USER_PROJECT_BASE_URL = "/dashboard/user-project";
 
@@ -102,13 +104,23 @@ export class UserFileSectionComponent {
   }
 
   public deleteUserFileEntry(userFileEntry: DashboardUserFileEntry): void {
-    this.userFileService.deleteDashboardUserFileEntry(userFileEntry).subscribe(
-      () => this.refreshDashboardFileEntries(),
-      (err: unknown) => {
-        // @ts-ignore // TODO: fix this with notification component
-        (err: unknown) => alert("Can't delete the file entry: " + err.error);
-      }
-    );
+    const modalRef = this.modalService.open(DeletePromptComponent);
+    modalRef.componentInstance.deletionType = "file";
+    modalRef.componentInstance.deletionName = userFileEntry.file.name;
+
+    from(modalRef.result)
+      .pipe(untilDestroyed(this))
+      .subscribe((confirmToDelete: boolean) => {
+        if (confirmToDelete && userFileEntry.file.fid !== undefined) {
+          this.userFileService.deleteDashboardUserFileEntry(userFileEntry).subscribe(
+            () => this.refreshDashboardFileEntries(),
+            (err: unknown) => {
+              // @ts-ignore // TODO: fix this with notification component
+              (err: unknown) => alert("Can't delete the file entry: " + err.error);
+            }
+          );
+        }
+      });
   }
 
   public disableAddButton(): boolean {
